test: move pipeline assets and pipelines into shared fixtures
