+++ test/test_helper.exs
@@ -1,5 +1,6 @@
 Code.require_file("support/fixtures/assets/basic_assets.ex", __DIR__)
 Code.require_file("support/fixtures/assets/graph_assets.ex", __DIR__)
+Code.require_file("support/fixtures/assets/pipeline_assets.ex", __DIR__)
 Code.require_file("support/fixtures/assets/runner_assets.ex", __DIR__)
 Code.require_file("support/favn_test_setup.ex", __DIR__)
